+++ src/lib/app/events.ts
@@ -1,6 +1,5 @@
 import type {EventInfo} from '$lib/vocab/event/event';
-import {LoginAccount, LogoutAccount} from '$lib/vocab/account/accountEvents';
-import {SetSession} from '$lib/session/sessionEvents';
+import {LoginAccount, LogoutAccount, SetSession} from '$lib/session/sessionEvents';
 import {
 	CreateCommunity,
 	ReadCommunity,
