+++ src/lib/vocab/account/account.schema.ts
@@ -12,12 +12,6 @@ export const AccountSchema = {
 	additionalProperties: false,
 };
 
-// TODO make the `create_account` event in accountEvents.ts
-export interface CreateAccountParams {
-	name: string;
-	password: string;
-}
-
 export const AccountModelSchema = {
 	$id: '/schemas/AccountModel.json',
 	type: 'object',
